% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myboot2}
\alias{myboot2}
\title{myBoot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{data to use}

\item{fun}{function to use}

\item{alpha}{interval deviation}

\item{cx}{int number}

\item{...}{extra parameters}
}
\value{
returns plot of the result
}
\description{
myBoot2
}
\examples{
myboot2(iter = 10000, x = 0:20, alpha = 0.05, fun = "mean", xlab = "mean/median", col = "Orange")
}
