% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Title}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{the sample size}

\item{iter}{number of iterations}
}
\value{
returns the histogram of the sums
}
\description{
Title
}
\examples{
myclt(n=5, iter=1000)
}
