% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{the function to be used}

\item{theta}{the variable to find the likelihood}
}
\value{
plot of the max likelihood and probability of a 'head'
}
\description{
mymaxlikg
}
\examples{
mymaxlikg(theta=seq(0,1,length=10000),
lfun=function(theta){log(dbinom(2,prob=theta,size=6)) + log(dbinom(4,prob=theta,size=10))})
}
