% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMultiFunct.R
\name{mymult}
\alias{mymult}
\title{Title}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{size}

\item{p}{probability}
}
\value{
a numeric vector
}
\description{
Title
}
\examples{
mymult(iter=100, n=10, p=c(1,1,1,1)/4)
}
