% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a, x = NULL)
}
\arguments{
\item{mu}{represents the mean}

\item{sigma}{the standard deviation}

\item{a}{probability endpoint, P(Y <= a)}

\item{x}{NULL}
}
\value{
a list containing mu, sigma, and area
}
\description{
Title
}
\examples{
myncurve(mu=4, sigma=2, a=6)
}
